front end redesign v1
This PR serves as a starting point for the frontend design. Some content is placeholder content and strictly visual, including most of the party bid section. Afterwards, state/logic will be dropped in as it becomes available

Attempted to delete as little code as possible. Until the frontend design is finalized and it’s determined old code no longer serves a purpose, commented code was left "as is"

Known bugs:
- Formatic wallet connect button crashes currently, have not looked into why
- State does not change on metamask wallet connect